requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pg8000
python-dotenv==1.0.0
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
email-validator
